Refactor: 패키지 구조 및 컨트롤러의 리턴 URI 리팩토링
- global 패키지의 구조 변경 (exception, util, entity, config 네 가지 하위 패키지로 분류)
- TodoController와 CommentController의 addTodo, addComment 메서드에서 리턴 URI 수정
(Spring.format을 사용하지 않고, URI 문자열에 $를 사용한 직접 대입)
