+++ src/main/kotlin/org/hunzz/todoapplication/global/entity/BaseEntity.kt
@@ -1,4 +1,4 @@
-package org.hunzz.todoapplication.global
+package org.hunzz.todoapplication.global.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.EntityListeners
